Removed some warnings. Some errors emerged.

+++ cg_framework/display.h
@@ -10,7 +10,7 @@ extern GLfloat lightIntensity[];
 extern GLfloat ambientComponent[];
 extern GLfloat diffuseColor[];
 extern bool exploding;
-extern double actualfragment;
+extern GLfloat actualfragment;
 extern GLfloat blink;
